% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fun.R
\name{extractgridded}
\alias{extractgridded}
\title{Extract Gridded Data from NetCDF Based on Shapefile}
\usage{
extractgridded(
  shapefile_path,
  netcdf_path,
  varname,
  state_field,
  district_field,
  fun = mean,
  na.rm = TRUE,
  output_csv = NULL
)
}
\arguments{
\item{shapefile_path}{The path to the shapefile containing region boundaries.}

\item{netcdf_path}{The path to the NetCDF file containing gridded data.}

\item{varname}{The name of the variable to extract from the NetCDF file.}

\item{state_field}{The column name in the shapefile that contains state or region names.}

\item{district_field}{The column name in the shapefile that contains district or subregion names.}

\item{fun}{The function to apply to the extracted data (default is mean).}

\item{na.rm}{Whether to remove NA values (default is TRUE).}

\item{output_csv}{Optional; path to save the results as a CSV file.}
}
\value{
A data frame with the extracted and aggregated data, and state/district information.
}
\description{
This function extracts data from a NetCDF file based on the regions defined in a shapefile.
The user provides the names of the columns in the shapefile that represent state and district.
The function can apply an aggregation function (e.g., mean, sum) to the extracted data.
}
